Adding cmake build

+++ checker_texture.h
@@ -17,7 +17,7 @@ private:
 
 vec3 checker_texture::value(float u, float v, const vec3 & p) const
 {
-    float sines = std::sinf(10 * p.x()) * std::sinf(10 * p.y()) * std::sinf(10 * p.z());
+    float sines = std::sin(10 * p.x()) * std::sin(10 * p.y()) * std::sin(10 * p.z());
     if (sines < 0)
     {
         return pOdd->value(u, v, p);
